components_ui.h has moved.
Chromium change:

https://chromium.googlesource.com/chromium/src/+/4ba6c9a620c7705767544273e9639ad48faeceda

commit 4ba6c9a620c7705767544273e9639ad48faeceda
Author: Demetrios Papadopoulos <dpapad@chromium.org>
Date:   Sat Dec 7 03:26:01 2019 +0000

    Cleanup: Move chrome://components C++ classes to their own subfolder.

    Bug: None
